requirements.txt
kivy==2.3.1
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
kivy_deps.sdl2_dev==0.8.0
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
gspread
openpyxl
jinja2